- Made all (relevant) code compile at Visual Studio 2012's warning level 4. - Changed type of st_ucon64_t::parport from int to uint16_t.

+++ trunk/ucon64/src/backup/cd64.c
@@ -212,7 +212,7 @@ cd64_init (void)
     }
 
 #ifndef USE_PPDEV
-  if (ucon64.parport == UCON64_UNKNOWN)
+  if (ucon64.parport == (uint16_t) UCON64_UNKNOWN)
     {
       fputs ("ERROR: No port or invalid port specified\n"
              "TIP:   Specify one with --port or in the configuration file\n", stderr);
