+++ trunk/ucon64/src/backup/smsgg-pro.h
@@ -30,10 +30,10 @@ Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
 extern const st_getopt2_t smsggpro_usage[];
 
 #ifdef USE_PARALLEL
-extern int smsgg_read_rom (const char *filename, unsigned int parport, int size);
-extern int smsgg_write_rom (const char *filename, unsigned int parport);
-extern int smsgg_read_sram (const char *filename, unsigned int parport, int start_bank);
-extern int smsgg_write_sram (const char *filename, unsigned int parport, int start_bank);
+extern int smsgg_read_rom (const char *filename, unsigned short parport, int size);
+extern int smsgg_write_rom (const char *filename, unsigned short parport);
+extern int smsgg_read_sram (const char *filename, unsigned short parport, int start_bank);
+extern int smsgg_write_sram (const char *filename, unsigned short parport, int start_bank);
 #endif
 
 #endif
