+++ trunk/ucon64/src/backup/tototek.h
@@ -22,7 +22,7 @@ Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
 #define TOTOTEK_H
 
 #ifdef USE_PARALLEL
-extern void ttt_init_io (unsigned int port);
+extern void ttt_init_io (unsigned short port);
 extern void ttt_deinit_io (void);
 extern void ttt_set_ai_data (unsigned char ai, unsigned char data);
 extern void ttt_rom_enable (void);
@@ -30,7 +30,7 @@ extern void ttt_rom_disable (void);
 extern void ttt_ram_enable (void);
 extern void ttt_ram_disable (void);
 extern void ttt_write_mem (int addr, unsigned char b);
-extern unsigned short int ttt_get_id (void);
+extern unsigned short ttt_get_id (void);
 extern void ttt_read_rom_b (int addr, unsigned char *buf);
 extern void ttt_read_rom_w (int addr, unsigned char *buf);
 extern void ttt_read_ram_b (int addr, unsigned char *buf);
